move zwlist_t into own file
