+++ gzw_helpers.cpp
@@ -45,46 +45,6 @@ void string_freeBytes(bytes_t *bytes)
     free(bytes);
 }
 
-/*
- *
- * C types and functions.
- *
- */
-
-zwlist_t *zwlist_new()
-{
-    std::vector<std::string> *vec = new std::vector<std::string>();
-    assert(vec);
-    return (zwlist_t*)vec;
-}
-
-int zwlist_size(zwlist_t *list)
-{
-    assert(list);
-    std::vector<std::string> *vec = (std::vector<std::string>*)list;
-    return vec->size();
-}
-
-char* zwlist_at(zwlist_t *list, int pos)
-{
-    assert(list);
-    std::vector<std::string> *vec = (std::vector<std::string>*)list;
-    return zwhelper_makeCString(vec->at(pos));
-}
-
-void zwlist_free(zwlist_t *list)
-{
-    assert(list);
-    std::vector<std::string> *vec = (std::vector<std::string>*)list;
-    delete vec;
-}
-
-/*
- *
- * C++ only functions.
- *
- */
-
 char* zwhelper_makeCString(std::string &str)
 {
     char *cstr = (char*)malloc(sizeof(char) * str.size()+1);
@@ -93,11 +53,4 @@ char* zwhelper_makeCString(std::string &str)
     	strncpy(cstr, str.c_str(), str.size());
     }
     return cstr;
-}
-
-zwlist_t* zwlist_copy(std::vector<std::string> &vec)
-{
-    std::vector<std::string> *out = new std::vector<std::string>(vec);
-    assert(out);
-    return (zwlist_t*)out;
 }
