+++ gzw_helpers.h
@@ -29,16 +29,6 @@ void string_setByteAt(bytes_t *bytes, uint8_t value, size_t position);
 uint8_t string_byteAt(bytes_t *bytes, size_t position);
 void string_freeBytes(bytes_t *bytes);
 
-/**
- * C friendly representation of std::vector<std::string>*.
- */
-typedef void* zwlist_t;
-
-zwlist_t* zwlist_new();
-int zwlist_size(zwlist_t *list);
-char* zwlist_at(zwlist_t *list, int pos);
-void zwlist_free(zwlist_t *list);
-
 #ifdef __cplusplus
 } // end extern "C"
 
@@ -48,15 +38,7 @@ void zwlist_free(zwlist_t *list);
  *
  */
 
-/**
- * Create and return a copy of the std::string as a char*. The returned pointer
- * must be freed!
- * @param  str The string to copy.
- * @return     The new C string.
- */
 char* zwhelper_makeCString(std::string &str);
-
-zwlist_t* zwlist_copy(std::vector<std::string> &vec);
 
 #endif
